Fix minor warnings.
When compiling with -Wall/-Werror, several warnings related to signed/unsigned
comparison and an incorrect format string kill the build. Additionally, when
compiling under GCC 4.8.x, `max_align_t` is not a member of `std`. This change
fixes these minor errors.

PiperOrigin-RevId: 284216526

+++ tensorflow/lite/experimental/micro/micro_interpreter.cc
@@ -21,7 +21,7 @@ limitations under the License.
 
 namespace tflite {
 namespace {
-const int kStackDataAllocatorSize = 128;
+const size_t kStackDataAllocatorSize = 128;
 class StackDataAllocator : public BuiltinDataAllocator {
  public:
   void* Allocate(size_t size) override {
@@ -91,7 +91,7 @@ MicroInterpreter::MicroInterpreter(const Model* model,
   // NOTE: This requires that the flatbuffer is held in memory which can be
   // modified by this process.
   if (!FLATBUFFERS_LITTLEENDIAN) {
-    for (int t = 0; t < tensors_size(); ++t) {
+    for (size_t t = 0; t < tensors_size(); ++t) {
       TfLiteTensor* thisTensor = &context_.tensors[t];
       if (thisTensor->allocation_type == kTfLiteMmapRo)
         CorrectTensorEndianness(thisTensor);
